Redesign of gui part 2. Not completed yet.

+++ MSDefragLib/Defragmenter/BaseDefragmenter.cs
@@ -12,7 +12,7 @@ namespace MSDefragLib.Defragmenter
         public abstract void Start(string parameter);
         public abstract void Stop(int timeoutMs);
 
-        public abstract int NumClusters { get; set; }
+        public abstract UInt64 NumClusters { get; set; }
 
         public abstract event ClustersModifiedHandler ClustersModified;
         public abstract event LogMessageHandler LogMessage;
